requirements.txt
python-dotenv>=1.0.0
pandas>=2.0.0
taiga-api>=1.1.0
holidays>=0.28
pytz>=2023.3
matplotlib>=3.7.0
seaborn>=0.12.0
openpyxl>=3.1.0
jupyter>=1.0.0
ipython>=8.0.0
streamlit>=1.20.0